视频点赞（未完）

+++ src/main/java/com/ligg/pojo/Comments.java
@@ -12,7 +12,7 @@ import java.time.LocalDateTime;
 public class Comments {
     private Integer id;//
     private Integer videoId;//视频id
-    private Integer userId;//用户id
+    private Long userId;//用户id
     private String userPic;//用户头像
     private String nickname;//用户昵称
     private String content;//评论内容
